jpa customer patch test

+++ src/main/java/sam/springframework/spring6restmvc/services/BeerServiceJPA.java
@@ -3,6 +3,7 @@ package sam.springframework.spring6restmvc.services;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 import sam.springframework.spring6restmvc.mappers.BeerMapper;
 import sam.springframework.spring6restmvc.model.BeerDTO;
 import sam.springframework.spring6restmvc.respositories.BeerRepository;
@@ -48,6 +49,7 @@ public class BeerServiceJPA implements BeerService {
             foundBeer.setBeerStyle(beer.getBeerStyle());
             foundBeer.setUpc(beer.getUpc());
             foundBeer.setPrice(beer.getPrice());
+            foundBeer.setQuantityOnHand(beer.getQuantityOnHand());
             atomicReference.set(Optional.of(
                     beerMapper.beerToBeerDto(beerRepository.save(foundBeer)))
             );
@@ -59,12 +61,38 @@ public class BeerServiceJPA implements BeerService {
     }
 
     @Override
-    public void deleteById(UUID beerId) {
-        beerRepository.deleteById(beerId);
+    public Boolean deleteById(UUID beerId) {
+        if(beerRepository.existsById(beerId)){
+            beerRepository.deleteById(beerId);
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public void patchBeerId(UUID beerId, BeerDTO beer) {
+    public Optional<BeerDTO> patchBeerId(UUID beerId, BeerDTO beer) {
+        AtomicReference<Optional<BeerDTO>> atomicReference = new AtomicReference<>();
 
+        beerRepository.findById(beerId).ifPresentOrElse(foundBeer -> {
+            if(StringUtils.hasText(beer.getBeerName())) {
+                foundBeer.setBeerName(beer.getBeerName());
+            }
+            if(beer.getBeerStyle() != null) {
+                foundBeer.setBeerStyle(beer.getBeerStyle());
+            }
+            if(StringUtils.hasText(beer.getUpc())) {
+                foundBeer.setUpc(beer.getUpc());
+            }
+            if(beer.getPrice() != null) {
+                foundBeer.setPrice(beer.getPrice());
+            }
+            atomicReference.set(Optional.of(
+                    beerMapper.beerToBeerDto(beerRepository.save(foundBeer)))
+            );
+        }, () -> {
+            atomicReference.set(Optional.empty());
+        });
+
+        return atomicReference.get();
     }
 }
